test: rename global instructions suite
