+++ codex-rs/core/tests/suite/global_instructions.rs
