+++ codex-rs/core/tests/suite/mod.rs
@@ -52,7 +52,7 @@ mod deprecation_notice;
 mod exec;
 mod exec_policy;
 mod fork_thread;
-mod global_instructions_regression;
+mod global_instructions;
 #[cfg(not(target_os = "windows"))]
 mod guardian_review;
 mod hierarchical_agents;
